WIP: Reduced size of logged pen data
When writing data to re2 (via /dev/input/event1), it is necessary to use the
codes 330 (PEN IN) and 330 (TOOL DOWN). This first selected to tool being used,
the second indicates whether it has touched the writing surface or not.

Similarly, when ending a stoke, need to use these codes with a '0' value.